lib: move assets JS API interface to lib/internal
So later can be used by both ZMQ Agent and the upcoming GRPC Agent.
