+++ test/parallel/test-bootstrap-modules.js
@@ -117,6 +117,7 @@ expected.atRunTime = new Set([
   'NativeModule internal/agents/statsd/lib/nsolid',
   'NativeModule internal/agents/zmq/lib/agent',
   'NativeModule internal/agents/zmq/lib/nsolid',
+  'NativeModule internal/nsolid_assets',
   'NativeModule internal/nsolid_loader',
   'NativeModule internal/nsolid_promise_tracking',
 ]);
